+++ WebAppGenerator/src/generator/MainGenerator.java
@@ -3,6 +3,8 @@ package generator;
 import java.io.File;
 import java.io.FileWriter;
 import java.io.Writer;
+import java.nio.file.Files;
+import java.nio.file.Paths;
 
 import freemarker.template.Template;
 import utils.ProjectInfo;
@@ -11,7 +13,6 @@ public class MainGenerator extends AbstractGenerator {
 
 	@Override
 	public void init() {
-
 		String projectPackage = projectInfo.getProjectPackage();
 		String projectName = projectInfo.getApplicationName();
 
@@ -21,24 +22,26 @@ public class MainGenerator extends AbstractGenerator {
 
 	@Override
 	public void generate() {
+		String mainPath = ProjectInfo.getInstance().getBaseHandwrittenFilesPath() + File.separatorChar
+				+ projectInfo.getApplicationName() + ".java";
+		
+		if (Files.notExists(Paths.get(mainPath))) {
+			try {
+				Template temp = generatorInfo.getConfiguration().getTemplate("main.ftl");
 
-		try {
-			Template temp = generatorInfo.getConfiguration().getTemplate("main.ftl");
+				File file = new File(mainPath);
+				file.createNewFile();
 
-			File file = new File(ProjectInfo.getInstance().getBasePath() + "//" + projectInfo.getApplicationName() + ".java");
-			file.createNewFile();
+				Writer fileWriter = new FileWriter(file);
 
-			Writer fileWriter = new FileWriter(file);
+				temp.process(model, fileWriter);
 
-			temp.process(model, fileWriter);
+				fileWriter.flush();
+				fileWriter.close();
 
-			fileWriter.flush();
-			fileWriter.close();
-
-		} catch (Exception e) {
-			e.printStackTrace();
+			} catch (Exception e) {
+				e.printStackTrace();
+			}
 		}
-
 	}
-
 }
